Rewrite image build (#14)
* spi: add Region

* images: add Inject to CloverImageService

* Rename Properties files

* Restructure - pass 1 - reduce inheritence - more composing

* Fix following rebase conflicts

* Don’t deprecate Area.of()

* Remove unused shadow left over from rebase

* Add Dimensions

* Get correct crop for kindle from paperback

* Use integers for pixels once rescaled scaled

* front and spine crops should be relative to wrapCrop

* Use front crop when placing text on front page

* add logging and minimum cover width at 1000

* Switch Makefile to issue 3

+++ spi/src/main/java/net/kemitix/clover/spi/images/Area.java
@@ -9,10 +9,10 @@ import lombok.ToString;
 @ToString
 public class Area {
 
-    private final int width;
-    private final int height;
+    private final float width;
+    private final float height;
 
-    public static Area of(final int width, final int height) {
+    public static Area of(final float width, final float height) {
         return Area.builder()
                 .width(width)
                 .height(height)
